Fixing tests.
