+++ src/components/AccountForm/AccountForm.test.tsx
@@ -3,9 +3,20 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import { axe, toHaveNoViolations } from "jest-axe";
 import { TestProviderWrapper } from "../../../testHelper/utils";
-import AccountForm from "./AccountForm";
+import AccountForm from ".";
 
 expect.extend(toHaveNoViolations);
+
+describe("AccountForm accessibility check", () => {
+  test("passes axe accessibility test", async () => {
+    const { container } = render(
+      <TestProviderWrapper>
+        <AccountForm />
+      </TestProviderWrapper>
+    );
+    expect(await axe(container)).toHaveNoViolations();
+  });
+});
 
 describe("AccountForm", () => {
   beforeEach(() => {
@@ -16,26 +27,12 @@ describe("AccountForm", () => {
     );
   });
 
-  test("renders username, pin, and home library code fields", () => {
+  test("renders username and pin fields", () => {
     expect(
       screen.getByRole("textbox", { name: "Username Required" })
     ).toBeInTheDocument();
     expect(
       screen.getByRole("textbox", { name: "PIN Required" })
     ).toBeInTheDocument();
-    expect(
-      screen.getByRole("textbox", { name: "Home Library:" })
-    ).toBeInTheDocument();
-  });
-});
-
-describe("AccountForm accessibility check", () => {
-  test("passes axe accessibility test", async () => {
-    const { container } = render(
-      <TestProviderWrapper>
-        <AccountForm />
-      </TestProviderWrapper>
-    );
-    expect(await axe(container)).toHaveNoViolations();
   });
 });
