+++ src/components/ConfirmationContainer/ConfirmationContainer.test.tsx
@@ -3,7 +3,8 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import { axe, toHaveNoViolations } from "jest-axe";
 import ConfirmationContainer from ".";
-import { FormResults } from "../../interfaces";
+import { FormResults, FormInputData } from "../../interfaces";
+import { FormDataContextProvider } from "../../context/FormDataContext";
 
 expect.extend(toHaveNoViolations);
 
@@ -16,16 +17,31 @@ const formResults: FormResults = {
   patronId: 1234567,
   name: "Tom Nook",
 };
+const formState = {
+  results: formResults,
+  errorObj: {},
+  csrfToken: "",
+  formValues: {} as FormInputData,
+  addressResponse: {},
+};
 
 describe("Confirmation", () => {
   test("passes axe accessibility test", async () => {
-    const { container } = render(<ConfirmationContainer />);
+    const { container } = render(
+      <FormDataContextProvider initState={formState}>
+        <ConfirmationContainer />
+      </FormDataContextProvider>
+    );
 
     expect(await axe(container)).toHaveNoViolations();
   });
 
   test("renders with the basic form submission result displaying", () => {
-    render(<ConfirmationContainer />);
+    render(
+      <FormDataContextProvider initState={formState}>
+        <ConfirmationContainer />
+      </FormDataContextProvider>
+    );
 
     expect(
       screen.getByText("Thank you for submitting your application.")
@@ -40,7 +56,11 @@ describe("Confirmation", () => {
   });
 
   test("renders the NYPL card info", () => {
-    render(<ConfirmationContainer />);
+    render(
+      <FormDataContextProvider initState={formState}>
+        <ConfirmationContainer />
+      </FormDataContextProvider>
+    );
 
     expect(screen.getByText("MEMBER NAME")).toBeInTheDocument();
     expect(screen.getByText("Tom Nook")).toBeInTheDocument();
